backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
python-multipart==0.0.9
passlib==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-telegram-bot==21.6
